Added Basic Notification manager
This works with an alarm quite well.  Moved vibrate right when the notif
happens.  Also changed test message to not use the notification manager
and launch service directly.

No need for a service to manager the alarms.  However a service will be
used to periodically check that the app is running, reset random times,
and other tasks.